requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
ruff>=0.4
mypy>=1.10
django-stubs>=5.0
coverage[toml]>=7.5
factory-boy>=3.3
